move CustomObject to top level class
